+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './Constants'
-export * from './DaemonExec'
-export * from './Daemon'
-export * from './Client'
-export * from './Http2Client'
-export * from './ZeroconfDaemon'
-export * from './ZeroconfClient'
+export * from './Daemon/DaemonExec'
+export * from './Daemon/Daemon'
+export * from './Client/Client'
+export * from './Client/Http2Client'
+export * from './Zeroconf/ZeroconfDaemon'
+export * from './Zeroconf/ZeroconfClient'
