+++ src/programs/server.ts
@@ -4,9 +4,9 @@ import * as bonjour from 'bonjour'
 import { AddressInfo } from 'net'
 import { hostname } from 'os'
 
-import { publishServer } from './ZeroconfDaemon'
-import { createDaemon } from './Daemon'
-import { SERVICE_TYPE } from './Constants'
+import { publishServer } from '../Zeroconf/ZeroconfDaemon'
+import { createDaemon } from '../Daemon/Daemon'
+import { SERVICE_TYPE } from '../Constants'
 
 const zeroconf = bonjour()
