+++ test/ZeroconfClient.test.ts
@@ -2,8 +2,8 @@ import { mock } from 'jest-mock-extended'
 import { Bonjour, Browser, RemoteService } from 'bonjour'
 import { EventEmitter } from 'events'
 
-import { supplyClient } from '../src/ZeroconfClient'
-import { Client } from '../src/Client'
+import { supplyClient } from '../src/Zeroconf/ZeroconfClient'
+import { Client } from '../src/Client/Client'
 import { SERVICE_TYPE } from '../src/Constants'
 
 const MOCK_SERVICE_HOST = 'example.com'
